Move workspace edit tests to separate file
